+++ src/pipeline/pipe.js
@@ -1,5 +1,6 @@
 import Pipeline from './Pipeline'
 
 export default function pipe (...middleware) {
-  return new Pipeline(middleware)
+  const pipeline = Pipeline.make(middleware)
+  return pipeline
 }