+++ src/server/http/responseDecorators/defaultDecorators.js
@@ -1,0 +1,9 @@
+import noUndefinedDecorator from './noUndefinedDecorator'
+import toStringDecorator from './toStringDecorator'
+import jsxDecorator from './jsxDecorator'
+
+export default [
+  toStringDecorator,
+  noUndefinedDecorator,
+  jsxDecorator
+]