+++ src/server/http/responseDecorators/toStringDecorator.js
@@ -1,0 +1,11 @@
+import Response from '../Response'
+
+export default function toStringDecorator (response) {
+  if (response instanceof Response) {
+    return response
+  }
+  if (typeof response === 'object') {
+    return Response.json(response)
+  }
+  return Response.ok(String(response))
+}